FEAT: record messages sent on ticket channel

+++ src/commands/discord/closeTicket.js
@@ -1,4 +1,4 @@
-import { ticketCategoryId, botId } from '../../../config.json'
+import { channels, botId } from '../../../config.json'
 import { PermissionFlagsBits } from "discord.js"
 import * as ticketService from '../../service/ticket'
 
@@ -9,6 +9,10 @@ module.exports = {
     // deleted: Boolean,
     callback: async (client, interaction) => {
         try{
+            if(interaction.channel.parentId != channels.ticketsCategory || interaction.channelId == channels.openTicket){
+                interaction.reply('Para finalizar um chamado, execute este comando no canal do chamado que deseja fechar!')
+                return
+            }
             ticketService.closeTicket({channelId: interaction.channelId})
             interaction.reply(`<@${interaction.user}> chamado finalizado!`)
             return await interaction.channel.delete()
